Improved plugin performance by more than 20x.
main.ts:

- markers are now stored in a Map for O(1) access. There is no need for
the findMarkerMatch function.
- The rectangle box information of each clickableElement is fetched in
first place and stored, so there's no need to query again the DOM when
creating the markers. This allows to avoid Layout Reflow penalties.
- clickableElements are now filtered, ditching elements out of workspace
bounds, saving a lot of work that will be done for non displayed
elements in the workspace.
- Remove the need to fetch markers rectangle box when updating markers.
It only needs to hide discarded markers on user input and update hints
on the remaining ones.

utils.ts:

- In updateMarkerText function remove childs more efficiently by setting
innerHTML to "".
- In createMarker function posX and posY are passed as parameters, no
need to fetch info from the DOM. No need to check if width and height
are 0. If there was an element with this settings it was filtered out.

utils.ts, types.ts and main.ts:
- Change text for hint to use the same term in all files.